First Application done second week exercises ongoing
First application that is going to be analysed throughout
the first six week is done. It should give us a sensible
example for most of the commonly used features of python.

It is sufficiently complex but simple enough to understand
what's going on once you have all the background due to
its minimal size.

Though builder paradigm is less used than mvc for example,
it maps well to the underlaying problem of different
formatted text objects.

The students need to be exposed to the parts of it as we
go along the parts of python.

The exercises for second week involve mostly io.

The topics that remain to be covered with exercises include:

- Writting from a list to a document
- writting binary documents.
- Problems with serialization of xml, json and why they should
not be read normally.
- Reading a file in chunks