Lecture done, exercices not done
Lecture is rather terse but should be sufficient for most cases.
Exercices are mostly about rewriting old solutions.
The idea is to make students feel that oop is more about viewing
your code in a different manner rather than using a different
syntax.

The solutions to exercice 03 and 04 are not done yet.